requirements.txt
langchain_groq
langchain_core
dotenv
langchain_community
langserve
api
uvicorn
Ollama